BasicPID op y-as nog steeds. als men de stick manueel laat afwijken, vermeerdert de tegenkracht met te tijd (vanwege de I component). domweg limiteren van de force is NOK: iwendig blijft de PID opbouwen, dus bij loslaten: overshoot!
BasicPID op y-as. schijnt goed te werken. we gebruiken nu material sliders. vertical sliders doen moeilijk. de hoogte is niet aanpasbaar, en wanneer er 2 naast mekaar staan, verkleinen de hoogtes van de sliders

+++ app/src/main/java/com/alaindef/state/UdpSender.kt
@@ -28,11 +28,11 @@ class UdpSender(private val ipAddress: String, private val portSender: Int) {
         return byteBuffer.array()
     }
 
-    fun sendUDP(forceX: Int, forceY: Int, IP: InetAddress, port: Int) {
+    fun sendUDP(forceX: Float, forceY: Float, IP: InetAddress, port: Int) {
         val policy = ThreadPolicy.Builder().permitAll().build()
         StrictMode.setThreadPolicy(policy)
 
-        val byteMessage = convertTheIndians(intArrayOf(0xAE, forceY, 0, forceX, 0, 0, 0, 0, 0))
+        val byteMessage = convertTheIndians(intArrayOf(0xAE, forceY.toInt(), 0, forceX.toInt(), 0, 0, 0, 0, 0))
         try {
             val socketS = DatagramSocket()
